+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - agent: Add `start_agent.sh` script to the repository
 - agent: Import default agent from the Pink balancer
-- control: Create `upkie.control` submodule
+- control: Create `upkie.controllers` submodule
 - control: Import MPC balancer class from the Pink balancer
 - cpp: Controller pipeline run by the spine after observers
 - envs: Add `Upkie-Servos-Mock` environment